+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add error skipping mechanism for some exceptions ([#271](https://github.com/SnowdogApps/magento2-menu/pull/271))
 - Fix Category Child node without parent category chosen makes menu disappear (DEV-99318)
 - Add required prop for category child field (DEV-99318)
+- Display a message indicating success or error after saving the menu. #274
 
 ## [2.21.0] - 2023-04-03
 ### Fixed